// Front-End-Rincon/ElRinconDeLasCartas/src/app/model/cardVendedorDto.ts
export class cardVendedorDto{
    CartaVendedorId: number;
    CartaFId: number;
    CartaFName: String;
    FImg: String;
    VendedorFId: number;
    Idioma: String;
    Estado: String;
    Cantidad: number;
    Precio: number;
}

export class cardVendedorConNombreDto{
    CartaVendedorId: number;
    CartaFId: number;
    CartaFName: String;
    FImg: String;
    VendedorFId: number;
    VendedorName : String;
    Idioma: String;
    Estado: String;
    Cantidad: number;
    Precio: number;
}